# blueprint/openapi.yaml
openapi: "3.0.3"
info:
  title: Saukele Wedding Registry API
  description: >
    Pool-funded wedding gift registry platform with kinship-tier gift
    suggestions, multi-currency support, and ACID-safe contributions.
  version: "1.0.0"
  contact:
    name: Saukele Support
    email: support@saukele.kz
servers:
  - url: http://localhost:3000/api
    description: Local development server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token. Obtain via POST /auth/login.
        Include as: Authorization: Bearer <token>
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        code:
          type: string
          nullable: true
          example: VALIDATION_ERROR
        error:
          type: string
          example: Resource not found
        message:
          type: string
          nullable: true
          example: Validation failed
        details:
          type: object
          nullable: true
        requestId:
          type: string
          nullable: true

    # ── Monetary precision-safe scalars ───────────────────────────────────
...
    UserPublic:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: asel@example.kz
        name:
          type: string
          example: Asel Nurlanovna
        role:
          type: string
          enum: [GUEST, REGISTRANT, ADMIN]
          example: REGISTRANT
        createdAt:
          type: string
          format: date-time
          example: "2026-01-15T10:00:00.000Z"
...
    RegisterRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
          example: asel@example.kz
        name:
          type: string
          example: Asel Nurlanovna
        password:
          type: string
          minLength: 8
          example: Qwerty123!
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: asel@example.kz
        password:
          type: string
          example: Qwerty123!
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: Login successful
        accessToken:
          type: string
          example: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/UserPublic'
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9...
    RefreshResponse:
      type: object
      properties:
        message:
          type: string
          example: Token refreshed successfully
        accessToken:
          type: string
          example: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9...

    # ── Registry ──────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Missing fields or password too short
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Password must be at least 8 characters
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Email already registered
        '422':
          description: Invalid email format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many registration attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Missing email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Invalid credentials
        '422':
          description: Invalid input format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts (5 per 15 min)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Access token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '400':
          description: Missing refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many refresh attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserPublic'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (blacklists token in Redis)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── REGISTRIES ────────────────────────────────────────────────────────────
...